Remove quasi-duplicate sync code using maybe-async-cfg
