+++ Cargo.toml
@@ -20,6 +20,7 @@ all-features = true
 
 [dependencies]
 defmt = { version = "0.3", optional = true }
+maybe-async-cfg = "0.2.3"
 
 [workspace]
 members = [
