+++ i2c/Cargo.toml
@@ -18,5 +18,10 @@ all-features = true
 
 [dependencies]
 embedded-hal = "1.0.0"
-embedded-hal-async = "1.0.0"
+embedded-hal-async = { version = "1.0.0", optional = true }
 display-interface = { version = "0.5.0", path = ".." }
+maybe-async-cfg = "0.2.3"
+
+[features]
+default = []
+async = ["embedded-hal-async"]
