+++ spi/src/lib.rs
@@ -1,31 +1,48 @@
-//! Generic SPI interface for display drivers
-
-#![no_std]
-
-mod asynch;
+//! Generic asynchronous SPI interface for display drivers
 
 use byte_slice_cast::*;
-use display_interface::{DataFormat, DisplayError, WriteOnlyDataCommand};
-use embedded_hal::{digital::OutputPin, spi::SpiDevice};
+#[cfg(feature = "async")]
+use display_interface::AsyncWriteOnlyDataCommand;
+#[cfg(not(feature = "async"))]
+use display_interface::WriteOnlyDataCommand;
+
+use display_interface::{DataFormat, DisplayError};
+use embedded_hal::digital::OutputPin;
+#[cfg(not(feature = "async"))]
+use embedded_hal::spi::SpiDevice;
+#[cfg(feature = "async")]
+use embedded_hal_async::spi::SpiDevice as AsyncSpiDevice;
 
 type Result = core::result::Result<(), DisplayError>;
-
 pub(crate) const BUFFER_SIZE: usize = 64;
 
-fn send_u8<SPI>(spi: &mut SPI, words: DataFormat<'_>) -> Result
+#[maybe_async_cfg::maybe(
+    sync(
+        cfg(not(feature = "async")),
+        keep_self,
+        idents(AsyncSpiDevice(sync = "SpiDevice"),)
+    ),
+    async(feature = "async", keep_self)
+)]
+async fn send_u8<SPI>(spi: &mut SPI, words: DataFormat<'_>) -> Result
 where
-    SPI: SpiDevice,
+    SPI: AsyncSpiDevice,
 {
     match words {
-        DataFormat::U8(slice) => spi.write(slice).map_err(|_| DisplayError::BusWriteError),
+        DataFormat::U8(slice) => spi
+            .write(slice)
+            .await
+            .map_err(|_| DisplayError::BusWriteError),
         DataFormat::U16(slice) => spi
             .write(slice.as_byte_slice())
+            .await
             .map_err(|_| DisplayError::BusWriteError),
         DataFormat::U16LE(slice) => {
             for v in slice.as_mut() {
                 *v = v.to_le();
             }
             spi.write(slice.as_byte_slice())
+                .await
                 .map_err(|_| DisplayError::BusWriteError)
         }
         DataFormat::U16BE(slice) => {
@@ -33,6 +50,7 @@ where
                 *v = v.to_be();
             }
             spi.write(slice.as_byte_slice())
+                .await
                 .map_err(|_| DisplayError::BusWriteError)
         }
         DataFormat::U8Iter(iter) => {
@@ -44,13 +62,16 @@ where
                 i += 1;
 
                 if i == buf.len() {
-                    spi.write(&buf).map_err(|_| DisplayError::BusWriteError)?;
+                    spi.write(&buf)
+                        .await
+                        .map_err(|_| DisplayError::BusWriteError)?;
                     i = 0;
                 }
             }
 
             if i > 0 {
                 spi.write(&buf[..i])
+                    .await
                     .map_err(|_| DisplayError::BusWriteError)?;
             }
 
@@ -66,6 +87,7 @@ where
 
                 if i == buf.len() {
                     spi.write(buf.as_byte_slice())
+                        .await
                         .map_err(|_| DisplayError::BusWriteError)?;
                     i = 0;
                 }
@@ -73,6 +95,7 @@ where
 
             if i > 0 {
                 spi.write(buf[..i].as_byte_slice())
+                    .await
                     .map_err(|_| DisplayError::BusWriteError)?;
             }
 
@@ -89,6 +112,7 @@ where
 
                 if i == len {
                     spi.write(buf.as_byte_slice())
+                        .await
                         .map_err(|_| DisplayError::BusWriteError)?;
                     i = 0;
                 }
@@ -96,6 +120,7 @@ where
 
             if i > 0 {
                 spi.write(buf[..i].as_byte_slice())
+                    .await
                     .map_err(|_| DisplayError::BusWriteError)?;
             }
 
@@ -126,24 +151,35 @@ impl<SPI, DC> SpiInterface<SPI, DC> {
     }
 }
 
-impl<SPI, DC> WriteOnlyDataCommand for SpiInterface<SPI, DC>
+#[maybe_async_cfg::maybe(
+    sync(
+        cfg(not(feature = "async")),
+        keep_self,
+        idents(
+            AsyncWriteOnlyDataCommand(sync = "WriteOnlyDataCommand"),
+            AsyncSpiDevice(sync = "SpiDevice"),
+        )
+    ),
+    async(feature = "async", keep_self)
+)]
+impl<SPI, DC> AsyncWriteOnlyDataCommand for SpiInterface<SPI, DC>
 where
-    SPI: SpiDevice,
+    SPI: AsyncSpiDevice,
     DC: OutputPin,
 {
-    fn send_commands(&mut self, cmds: DataFormat<'_>) -> Result {
+    async fn send_commands(&mut self, cmds: DataFormat<'_>) -> Result {
         // 1 = data, 0 = command
         self.dc.set_low().map_err(|_| DisplayError::DCError)?;
 
         // Send words over SPI
-        send_u8(&mut self.spi, cmds)
+        send_u8(&mut self.spi, cmds).await
     }
 
-    fn send_data(&mut self, buf: DataFormat<'_>) -> Result {
+    async fn send_data(&mut self, buf: DataFormat<'_>) -> Result {
         // 1 = data, 0 = command
         self.dc.set_high().map_err(|_| DisplayError::DCError)?;
 
         // Send words over SPI
-        send_u8(&mut self.spi, buf)
+        send_u8(&mut self.spi, buf).await
     }
 }
